refactor: restructure storage UI for better organization and UX
- Move non-route components from routes/storage to components/storage
- Implement proper route structure with sidebar layout pattern
- Add auto-navigation to first container on page load
- Update sidebar to show only container name and file count
- Move Upload File button to page header
- Fix file upload error by including full_file_name field
- Remove unnecessary container description banner from file list
- Clean up empty state handling in container view
