+++ web/.env.sample
@@ -8,3 +8,4 @@ VITE_FLX_INTERNAL_URL="http://${API_CONTAINER_NAME}:8080" # when running in dock
 VITE_FLX_API_URL="${API_URL}"
 VITE_FLX_BASE_DOMAIN="${BASE_DOMAIN}"
 VITE_FLX_HTTP_SCHEME=http
+
