+++ web/app/components/storage/index.ts
@@ -1,0 +1,7 @@
+export { StorageSidebar } from "./sidebar";
+export { ContainerList } from "./container-list";
+export { FileList } from "./file-list";
+export { CreateContainerDialog } from "./create-container-dialog";
+export { UpdateContainerDialog } from "./update-container-dialog";
+export { RenameFileDialog } from "./rename-file-dialog";
+export { FileUploadDialog } from "./file-upload-dialog";
