+++ web/app/routes.ts
@@ -26,7 +26,10 @@ export default [
       ]),
       routeFolder("functions", "./routes/functions/"),
       routeFolder("logs", "./routes/logs/"),
-      routeFolder("storage", "./routes/storage/"),
+      route("storage", "./routes/storage/sidebar.tsx", [
+        index("./routes/storage/index.tsx"),
+        route(":containerId", "./routes/storage/$containerId.tsx"),
+      ]),
       routeFolder("forms", "./routes/forms/"),
       routeFolder("backups", "./routes/backups/"),
     ]),
